made importConversation a real hook
